# schemas/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "INZBC API",
    "summary": "Trade intelligence and FTA services for the India New Zealand Business Council.",
    "version": "0.1.0"
  },
  "paths": {
    "/api/runs": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_api_runs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunOut"
                  },
                  "type": "array",
                  "title": "Response List Runs Api Runs Get"
                }
              }
            }
          },
          "401": {
            "description": "No session, or the session has expired or been idle too long."
          },
          "403": {
            "description": "Authenticated and refused: the CSRF token is missing or wrong, the account is no longer active, or the caller does not hold a role permitted this operation."
          }
        },
        "security": [
          {
            "SessionCookie": []
          }
        ]
      },
      "post": {
        "tags": [
          "Runs"
        ],
        "summary": "Create Run",
        "operationId": "create_run_api_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRunIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            }
          },
          "401": {
            "description": "No session, or the session has expired or been idle too long."
          },
          "403": {
            "description": "Authenticated and refused: the CSRF token is missing or wrong, the account is no longer active, or the caller does not hold a role permitted this operation."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SessionCookie": [],
            "CsrfToken": []
          }
        ]
      }
    },
    "/api/runs/{run_id}": {
      "get": {
        "tags": [
          "Runs"
        ],
        "summary": "Get Run",
        "operationId": "get_run_api_runs__run_id__get",
        "security": [
          {
            "SessionCookie": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            }
          },
          "401": {
            "description": "No session, or the session has expired or been idle too long."
          },
          "403": {
            "description": "Authenticated and refused: the CSRF token is missing or wrong, the account is no longer active, or the caller does not hold a role permitted this operation."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/start": {
      "post": {
        "tags": [
          "Runs"
        ],
        "summary": "Start Run",
        "description": "Draft -> Run Authorised. Launch authority; human gated (run-level, #227).",
        "operationId": "start_run_api_runs__run_id__start_post",
        "security": [
          {
            "SessionCookie": [],
            "CsrfToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            }
          },
          "401": {
            "description": "No session, or the session has expired or been idle too long."
          },
          "403": {
            "description": "Authenticated and refused: the CSRF token is missing or wrong, the account is no longer active, or the caller does not hold a role permitted this operation."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/pause": {
      "post": {
        "tags": [
          "Runs"
        ],
        "summary": "Pause Run",
        "description": "Awaiting CEO Decision -> Paused. CEO decision; `approval_ref` must name a\n`decision_records` row.",
        "operationId": "pause_run_api_runs__run_id__pause_post",
        "security": [
          {
            "SessionCookie": [],
            "CsrfToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            }
          },
          "401": {
            "description": "No session, or the session has expired or been idle too long."
          },
          "403": {
            "description": "Authenticated and refused: the CSRF token is missing or wrong, the account is no longer active, or the caller does not hold a role permitted this operation."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/resume": {
      "post": {
        "tags": [
          "Runs"
        ],
        "summary": "Resume Run",
        "description": "Paused -> Coverage Locked. Resumption authority; human gated (run-level, #227).",
        "operationId": "resume_run_api_runs__run_id__resume_post",
        "security": [
          {
            "SessionCookie": [],
            "CsrfToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            }
          },
          "401": {
            "description": "No session, or the session has expired or been idle too long."
          },
          "403": {
            "description": "Authenticated and refused: the CSRF token is missing or wrong, the account is no longer active, or the caller does not hold a role permitted this operation."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/runs/{run_id}/complete": {
      "post": {
        "tags": [
          "Runs"
        ],
        "summary": "Complete Run",
        "description": "Distributed -> Closed. Mechanical closeout; not human gated.",
        "operationId": "complete_run_api_runs__run_id__complete_post",
        "security": [
          {
            "SessionCookie": [],
            "CsrfToken": []
          }
        ],
        "parameters": [
          {
            "name": "run_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Run Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransitionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunOut"
                }
              }
            }
          },
          "401": {
            "description": "No session, or the session has expired or been idle too long."
          },
          "403": {
            "description": "Authenticated and refused: the CSRF token is missing or wrong, the account is no longer active, or the caller does not hold a role permitted this operation."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateRunIn": {
        "properties": {
          "run_number": {
            "type": "string",
            "title": "Run Number"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version"
          },
          "coverage_start_utc": {
            "type": "string",
            "title": "Coverage Start Utc"
          },
          "coverage_end_utc": {
            "type": "string",
            "title": "Coverage End Utc"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "run_number",
          "prompt_version",
          "coverage_start_utc",
          "coverage_end_utc"
        ],
        "title": "CreateRunIn"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "RunOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "run_number": {
            "type": "string",
            "title": "Run Number"
          },
          "state": {
            "type": "string",
            "title": "State"
          },
          "version": {
            "type": "integer",
            "title": "Version"
          },
          "prompt_version": {
            "type": "string",
            "title": "Prompt Version"
          },
          "coverage_start_utc": {
            "type": "string",
            "title": "Coverage Start Utc"
          },
          "coverage_end_utc": {
            "type": "string",
            "title": "Coverage End Utc"
          },
          "initiated_by": {
            "type": "string",
            "title": "Initiated By"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "id",
          "run_number",
          "state",
          "version",
          "prompt_version",
          "coverage_start_utc",
          "coverage_end_utc",
          "initiated_by"
        ],
        "title": "RunOut"
      },
...
      "TransitionIn": {
        "properties": {
          "expected_version": {
            "type": "integer",
            "title": "Expected Version"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "approval_ref": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Approval Ref"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "expected_version",
          "reason"
        ],
        "title": "TransitionIn",
        "description": "Body shared by all four lifecycle commands. `approval_ref` is optional at the schema level\nbecause only three of the four transitions are human-gated (`complete` is not) - `RunRepository`\nis still the sole enforcer of which ones actually require it; this model does not duplicate\nthat rule."
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "SessionCookie": {
        "type": "apiKey",
        "description": "Opaque server-side session, issued out of band. HttpOnly, Secure, SameSite=Lax and host-only, so it is not readable by script and does not cross an origin. Only its SHA-256 digest is stored.",
        "in": "cookie",
        "name": "inzbc_session"
      },
      "CsrfToken": {
        "type": "apiKey",
        "description": "Double-submit token, returned by GET /api/session. Required on every state-changing request: SameSite=Lax still permits a top-level cross-site POST, which is the shape of a form-submission CSRF.",
        "in": "header",
        "name": "X-CSRF-Token"
      }
    }
  }
}